+++ about_crawling.txt
@@ -1,0 +1,34 @@
+덕성여자대학교 컴퓨터공학전공 김지민입니다.
+
+제가 만든 crawling 폴더와 하위 파일에 대한 설명 파일 입니다.
+
+--------------------------------------------------------------------------------------
+
+* crawling 작동을 위해 꼭 있어야 하는 파일:
+    - crawling.py
+    - chromedriver.exe
+    - California_zipcodes_LosAngles.xlsx 
+
+* 각 파일 설명     
+    - crawling.py
+    제가 만든 크롤링 소프트웨어 입니다. selenium을 이용했으며, 핵심 소스 파일입니다.
+    !!! 절대 삭제하지 마세요 !!! 
+
+    주요 기능은 아래와 같습니다.
+        - 웹 브라우저 구글맵을 자동으로 켜기 
+        - California_zipcodes_LosAngles.xlsx 파일을 활용한 검색
+        - 사용자가 html에서 입력한 정보를 통해서 자동 검색어 입력 및 검색(search)
+        - 해당 정보에 대한 데이터 크롤링
+            - 가게 이름
+            - 위치
+            - 별점
+            - 리뷰 수
+            - dollar
+        - 자동 데이터 수집 및 저장
+        - 해당 데이터에 대한 csv 파일 생성
+
+    - chromedriver.exe
+    crawling.py 파일에서 필요로 하는 프로그램 파일입니다. 
+
+    - California_zipcodes_LosAngles.xlsx 
+    crawling.py을 이용한 검색을 진행할 때 필요합니다. 다른 검색을 하고 싶다면 해당 파일을 변경하면 됩니다.
