+++ savior/templates/main.html
@@ -39,5 +39,9 @@
 
     <a href="{% url 'savior:japan_pricelist' %}">nav바 시세 선택 페이지</a><br>
 
+
+    <!--* 식당 추천 페이지  -->
+    <a href="{% url 'savior:recommend' %}">식당 추천 페이지</a><br><br><br>
+
 </body>
 </html>
