+++ savior/templates/recommend_restaurant.html
@@ -1,0 +1,47 @@
+<!DOCTYPE html>
+<html>
+<head>
+    <title>식당 소개 페이지 </title>
+
+</head>
+<body>
+    <h1>Restaurant Recommendation</h1>
+    <form method="post" action="{% url 'savior:recommend' %}">
+        {% csrf_token %}
+        <label for="user_input_search">입력받는 변수 이름 Suser_input_search 입니다. 이를 작성하세요:</label> 
+
+
+        <input type="text" id="user_input_search" name="user_input_search" required>
+        <button type="submit">Search</button> </br></br></br>
+
+        <h2>검색 결과를 표로 보여드립니다 :</h2>
+        <h1>검색 시간이 조금 걸려요~ 기다려 주세용~ </h1>
+        <table>
+            <thead>
+                <tr>
+                    <th>food_name</th>
+                    <th>codezip</th>
+                    <th>rating</th>
+                    <th>reviews</th>
+                    <th>dollar</th>
+                    <th>address</th>
+                </tr>
+            </thead>
+            <tbody>
+                <!--* csv_data는 view에서 정의된 리스트  -->
+                {% for row in csv_data %} 
+                <tr>
+                    <td>{{ row.food_name }}</td>
+                    <td>{{ row.codezip }}</td>
+                    <td>{{ row.rating }}</td>
+                    <td>{{ row.reviews }}</td>
+                    <td>{{ row.dollar }}</td>
+                    <td>{{ row.address }}</td>
+                </tr>
+                {% endfor %}
+            </tbody>
+        </table>
+
+    </form>
+</body>
+</html>
